// src/Types/AccountType.ts
export enum AccountType {
    CLIENT = 'Client',
    PROFESSIONAL = 'Professional',
    BOTH = 'Both'
}

export enum MembershipOption {
    SUBSCRIPTION = 'Subscription',
    SINGLE = 'Single'
}

export enum ProfessionalServices {
    TREE_REMOVAL = "Tree Removal",
    ROOF_CLEANING = "Roof Cleaning",
    FENCE_INSTALLATION = "Fence Installation",
    OVEN_REPAIRS = "Oven Repairs",
    PLUMBING = "Plumbing"
}